backend/app/services/inference/depth_estimation.py
import asyncio
from typing import Dict, Any, Optional
from PIL import Image
import numpy as np
import logging

from app.config import settings

logger = logging.getLogger(__name__)

class DepthEstimationService:
    def __init__(self):
        self.models = {}
        self.processors = {}
        self.initialized = False
    
    async def initialize(self):
        """Initialize depth estimation models"""
        try:
            logger.info("Initializing depth estimation models...")
            
            model_name = "damo/cv_resnet_depth-estimation"
            logger.info(f"Loading model: {model_name}")
            
            # Mock model loading
            self.models[model_name] = "mock_depth_model"
            self.processors[model_name] = "mock_depth_processor"
            
            self.initialized = True
            logger.info("Depth estimation service initialized successfully")
            
        except Exception as e:
            logger.error(f"Failed to initialize depth estimation service: {e}")
            raise
    
    async def estimate_depth(
        self,
        image: Image.Image,
        model_name: str = "damo/cv_resnet_depth-estimation"
    ) -> Dict[str, Any]:
        """Estimate depth from image"""
        
        if not self.initialized:
            await self.initialize()
        
        try:
            # Mock depth estimation results
            depth_map = np.random.rand(image.height, image.width) * 100  # Mock depth values
            
            return {
                "depth_map": depth_map.tolist(),
                "min_depth": float(np.min(depth_map)),
                "max_depth": float(np.max(depth_map)),
                "mean_depth": float(np.mean(depth_map)),
                "model_used": model_name,
                "image_dimensions": {
                    "width": image.width,
                    "height": image.height
                }
            }
            
        except Exception as e:
            logger.error(f"Depth estimation failed: {e}")
            raise
    
    def _run_inference(self, image: Image.Image, model_name: str) -> np.ndarray:
        """Run actual depth estimation inference (placeholder)"""
        
        # This is where you would implement actual model inference
        # For now, returning mock depth map
        
        return np.random.rand(image.height, image.width) * 100
