backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
transformers==4.36.0
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
websockets==12.0
